Move Workflow model to workflows submodule
